Change folder structure for lambda related code
