+++ referral-form-data-process/handler.ts
@@ -1,5 +1,3 @@
 import { SQSEvent } from "aws-lambda";
 
-export const generateGoogleDoc = async (sqsEvent: SQSEvent) => {
-
-}
+export const generateGoogleDoc = async (sqsEvent: SQSEvent) => {};
