+++ referral-form-data-process/lib/generateGoogleAuth.ts
